requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.11
# To update, run:
#
#    pip-compile --output-file=dev.txt dev.in
#
asgiref==3.5.2
    # via
    #   -r base.txt
    #   django
asttokens==2.0.8
    # via
    #   -r base.txt
    #   stack-data
attrs==22.1.0
    # via pytest
backcall==0.2.0
    # via
    #   -r base.txt
    #   ipython
black==22.10.0
    # via -r dev.in
click==8.1.3
    # via black
decorator==5.1.1
    # via
    #   -r base.txt
    #   ipython
django==4.1.2
    # via
    #   -r base.txt
    #   django-admin-display
    #   django-debug-toolbar
    #   django-extensions
    #   django-jazzmin
    #   django-stubs
    #   djangorestframework
django-admin-display==1.3.0
    # via -r base.txt
django-debug-toolbar==3.7.0
    # via -r dev.in
django-extensions==3.2.1
    # via -r base.txt
django-jazzmin==2.5.0
    # via -r base.txt
django-json-widget==1.1.1
    # via -r base.txt
django-stubs==1.12.0
    # via -r dev.in
django-stubs-ext==0.5.0
    # via django-stubs
djangorestframework==3.14.0
    # via -r base.txt
environs==9.5.0
    # via -r base.txt
execnet==1.9.0
    # via pytest-xdist
executing==1.1.1
    # via
    #   -r base.txt
    #   stack-data
flake8==5.0.4
    # via -r dev.in
future==0.18.2
    # via
    #   -r base.txt
    #   django-json-widget
iniconfig==1.1.1
    # via pytest
ipython==8.5.0
    # via -r base.txt
isort==5.10.1
    # via -r dev.in
jedi==0.18.1
    # via
    #   -r base.txt
    #   ipython
marshmallow==3.18.0
    # via
    #   -r base.txt
    #   environs
matplotlib-inline==0.1.6
    # via
    #   -r base.txt
    #   ipython
mccabe==0.7.0
    # via flake8
mypy==0.982
    # via
    #   -r dev.in
    #   django-stubs
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.3
    # via
    #   -r base.txt
    #   marshmallow
    #   pytest
    #   pytest-sugar
paho-mqtt==1.6.1
    # via -r base.txt
parso==0.8.3
    # via
    #   -r base.txt
    #   jedi
pathspec==0.10.1
    # via black
pexpect==4.8.0
    # via
    #   -r base.txt
    #   ipython
pickleshare==0.7.5
    # via
    #   -r base.txt
    #   ipython
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.31
    # via
    #   -r base.txt
    #   ipython
psycopg2-binary==2.9.5
    # via -r base.txt
ptyprocess==0.7.0
    # via
    #   -r base.txt
    #   pexpect
pure-eval==0.2.2
    # via
    #   -r base.txt
    #   stack-data
pycodestyle==2.9.1
    # via flake8
pyflakes==2.5.0
    # via flake8
pygments==2.13.0
    # via
    #   -r base.txt
    #   ipython
pyparsing==3.0.9
    # via
    #   -r base.txt
    #   packaging
pytest==7.2.0
    # via
    #   -r dev.in
    #   pytest-django
    #   pytest-sugar
    #   pytest-xdist
pytest-django==4.5.2
    # via -r dev.in
<<<<<<< HEAD
pytest-sugar==0.9.6
=======
pytest-sugar==0.9.5
>>>>>>> 21.2-sw-integracja-broker-backend
    # via -r dev.in
pytest-xdist==3.0.2
    # via -r dev.in
python-dotenv==0.21.0
    # via
    #   -r base.txt
    #   environs
pytz==2022.5
    # via
    #   -r base.txt
    #   djangorestframework
<<<<<<< HEAD
pyyaml==6.0
    # via -r base.txt
=======
>>>>>>> 21.2-sw-integracja-broker-backend
six==1.16.0
    # via
    #   -r base.txt
    #   asttokens
sqlparse==0.4.3
    # via
    #   -r base.txt
    #   django
    #   django-debug-toolbar
stack-data==0.5.1
    # via
    #   -r base.txt
    #   ipython
termcolor==2.0.1
    # via pytest-sugar
tomli==2.0.1
    # via django-stubs
traitlets==5.5.0
    # via
    #   -r base.txt
    #   ipython
    #   matplotlib-inline
types-pytz==2022.5.0.0
    # via django-stubs
types-pyyaml==6.0.12.1
    # via django-stubs
typing-extensions==4.4.0
    # via
    #   django-stubs
    #   mypy
wcwidth==0.2.5
    # via
    #   -r base.txt
    #   prompt-toolkit
